gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
httpx==0.27.0
celery==5.4.0
redis==5.0.0
google-auth==2.22.0
requests==2.31.0
boto3==1.34.162
